+++ Changelog.md
@@ -5,6 +5,10 @@ v1.0.4 Unreleased
 * Propagate uncertainties from radiometry to OC products
 * Revisit question of overly aggressive deglitching
 
+2020-12-01:
+* Create default SeaBASS header file when new Config file is created.
+* Check the L2 OC products were actually generated before plotting.
+
 2020-11-30: 
 * Change L1E angular extrapolations to use constant values rather than linear extrapolation except for SOLAR_AZ and SZA.